Moved PluginManifest to Disa.Framework.Telegram.Mobile
When deploying a plugin to Android, the file path of what to copy over
(i.e, the [filePath] in 'deploy-register') should always point to the
assembly with the most references. In this case, the Telegram Plugin
Settings assembly, Disa.Framework.Telegram.Mobile encompasses all the
references required for the plugin to function.