Started implementing the DC manager
The DC manager will be in charge of keeping connections to other DCs
(other than the primary one) active and tearing them down when they're
unused.
At the moment, a connection is reauthed every time the Disa app is
restarted. Once completely tested, code will be written to cache the
auth on the device's disk so constant reauth won't be necessary.