Temporary hacked-fix for authenticator
NewNonceHash from server sometimes differs from what’s generated at the
client end. Not exactly sure why this is — for now, just rerun the
setup until it’s setup successfully.
Note: we can’t just remove the no-once check as the AuthKey generated
when the NewNonceHash is not-matching is invalid.