Bumped to Framework 13
This update removes the framework’s dependency on using
Xamarin.Form.Labs, fixes a possibility in an incorrect (but correct)
phone number comparison, and introduces the ability to grab an area
code from a number.

Since there is a deprecation in this release, using this framework
version requires all plugins to have a minimumFrameworkVersion of 13.